chore : 컴포넌트 분리 (불필요한 리렌더링 방지)
