+++ client/src/pages/search/Style.tsx
@@ -18,18 +18,6 @@ export const TitleContainer = styled.div`
   justify-content: space-between;
 `;
 
-export const RadioContainer = styled.fieldset`
-  border: 1px solid ${colors.gray3};
-  border-radius: 8px;
-  padding: 4px 6px;
-
-  color: ${colors.gray4};
-  font-size: 14px;
-
-  display: flex;
-  align-items: center;
-`;
-
 export const Items = styled.div`
   margin-top: 20px;
